pyproject.toml
[tool.poetry]
name = "worm_tracker"
version = "0.1.0"
description = "worm tracker with opencv"
authors = ["valerybriz <valerybriz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.12.0"
opencv-contrib-python = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
